#29 - Rename: 폴더 정리
