--- src/components/FormWrap.style.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const FormWrap = styled.form`
-	max-width: 55rem;
-	margin: 3rem auto 0;
-	@media (max-width: 390px) {
-		max-width: 32rem;
-	}
-`;
-
-export default FormWrap;
